refactor(gateway): split chat handlers by workflow
